moved handlebars helper classes.
