+++ src/main/java/io/swagger/codegen/languages/DefaultCodegenConfig.java
@@ -15,13 +15,13 @@ import io.swagger.codegen.v3.CodegenProperty;
 import io.swagger.codegen.v3.CodegenResponse;
 import io.swagger.codegen.v3.CodegenSecurity;
 import io.swagger.codegen.v3.SupportingFile;
-import io.swagger.codegen.handlebars.helpers.BaseItemsHelper;
-import io.swagger.codegen.handlebars.helpers.BracesHelper;
-import io.swagger.codegen.handlebars.helpers.HasHelper;
-import io.swagger.codegen.handlebars.helpers.HasNotHelper;
-import io.swagger.codegen.handlebars.helpers.IsHelper;
-import io.swagger.codegen.handlebars.helpers.IsNotHelper;
-import io.swagger.codegen.handlebars.helpers.StringUtilHelper;
+import io.swagger.codegen.v3.generators.handlebars.BaseItemsHelper;
+import io.swagger.codegen.v3.generators.handlebars.BracesHelper;
+import io.swagger.codegen.v3.generators.handlebars.HasHelper;
+import io.swagger.codegen.v3.generators.handlebars.HasNotHelper;
+import io.swagger.codegen.v3.generators.handlebars.IsHelper;
+import io.swagger.codegen.v3.generators.handlebars.IsNotHelper;
+import io.swagger.codegen.v3.generators.handlebars.StringUtilHelper;
 import io.swagger.codegen.v3.utils.ModelUtils;
 import io.swagger.v3.core.util.Json;
 import io.swagger.v3.core.util.Yaml;
@@ -90,7 +90,7 @@ import static io.swagger.codegen.v3.CodegenConstants.HAS_REQUIRED_EXT_NAME;
 import static io.swagger.codegen.v3.CodegenConstants.IS_ARRAY_MODEL_EXT_NAME;
 import static io.swagger.codegen.v3.CodegenConstants.IS_CONTAINER_EXT_NAME;
 import static io.swagger.codegen.v3.CodegenConstants.IS_ENUM_EXT_NAME;
-import static io.swagger.codegen.handlebars.helpers.ExtensionHelper.getBooleanValue;
+import static io.swagger.codegen.v3.generators.handlebars.ExtensionHelper.getBooleanValue;
 import static io.swagger.codegen.languages.CodegenHelper.getDefaultIncludes;
 import static io.swagger.codegen.languages.CodegenHelper.getImportMappings;
 import static io.swagger.codegen.languages.CodegenHelper.getTypeMappings;
