+++ src/main/java/io/swagger/codegen/languages/java/JavaInflectorServerCodegen.java
@@ -23,7 +23,7 @@ import java.util.Map;
 
 import static io.swagger.codegen.v3.CodegenConstants.HAS_ENUMS_EXT_NAME;
 import static io.swagger.codegen.v3.CodegenConstants.IS_ENUM_EXT_NAME;
-import static io.swagger.codegen.handlebars.helpers.ExtensionHelper.getBooleanValue;
+import static io.swagger.codegen.v3.generators.handlebars.ExtensionHelper.getBooleanValue;
 
 public class JavaInflectorServerCodegen extends AbstractJavaCodegen {
